chore: [#1413] deduplicate npm integration plugins
Vite, esbuild, and register all hand-rolled their own copies of compileFloe,
findProjectRoot, and the .floe/ cache reader despite @floeorg/core existing.

- Adds findProjectRoot to @floeorg/core (using register's algorithm — node_modules
  first, falls back to package.json, then to start dir; matches Floe's compiler).
- Adds @floeorg/core as a dependency of vite-plugin, esbuild-plugin, register.
- Removes duplicated readCompiledOutput / compileFloe / findProjectRoot from
  vite-plugin, esbuild-plugin, register.
- Drops the dead _source param from vite-plugin's call site.
