+++ integrations/esbuild-plugin/package.json
@@ -34,6 +34,9 @@
   "peerDependencies": {
     "esbuild": ">=0.17.0"
   },
+  "dependencies": {
+    "@floeorg/core": "workspace:*"
+  },
   "devDependencies": {
     "@types/node": "^25.5.2",
     "esbuild": "^0.27.3",
